Create "header" of webpage
Items of "left-links" box of "header" box are placed at top-left corner of the page.

Items of "right-links" box of "header" box are placed at top-right corner of page.

Space is allocated between list and corners of page to improve user experience.

Color of links are changed to improve readability.