Add changes to "sidebar" box
Color of links are set to white to improve readability.

Line under links are removed to improve readability.

List is placed at top-left corner of sidebar to enhance user experience.

Sidebar is set at max-width of 300px is improve user experience.

Shrinking ability of sidebar is set to 0 to enhance user experience.